+++ src/main/java/com/hseapple/dao/RoleEntity.java
@@ -7,11 +7,11 @@ import java.time.LocalDateTime;
 @Entity
 @Table(name = "user_role", schema = "public")
 public class RoleEntity {
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -25,7 +25,7 @@ public class RoleEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @NotNull
     @Column(name = "role_user")
